refactor: register 함수 삭제
- setValues와 겹치는 역할을 하므로 삭제
- htmlFor 삭제
